eternal_roll/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
graphviz==0.20.3
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
pydot==2.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
urllib3==2.2.1
